Edit CommentsController to handle edge cases

+++ app/controllers/comments_controller.rb
@@ -2,14 +2,14 @@ class CommentsController < ApplicationController
   before_action :authenticate_user!
 
   def create
-    @post = Post.find_by_id(params[:comment][:post_id])
+    @post = Post.find_by_id(params.dig(:comment).dig(:post_id))
 
     if @post.nil?
       flash[:alert] = 'Post does not exist.'
       return redirect_to root_path
     end
 
-    @comment = Comment.new(body: params[:comment][:body], user: current_user, post: @post)
+    @comment = Comment.new(body: params.dig(:comment).dig(:body), user: current_user, post: @post)
     
     if @comment.save
       redirect_to post_path(@post)
@@ -20,6 +20,11 @@ class CommentsController < ApplicationController
 
   def destroy
     comment = Comment.find_by_id(params[:id])
+
+    if comment.nil?
+      flash[:alert] = "Comment with id '#{params[:id]}' does not exist."
+      return redirect_to root_path
+    end
 
     if current_user == comment.user
       comment.destroy
